chore: move router from main.ts -> router.ts
